refactor(phase1): decouple shared/layout from landing-v2 + legacy app/pages
- useGoogleAuth: landing-v2/utils -> shared/hooks
- TopNav, Footer: landing-v2/components -> components/layout
- curatedListsConfig -> shared/lib/curatedLists
- CreateListModal, AddToListModal: app/pages/lists -> features/lists-v2
Rewired 21 importers (alias) + 2 legacy relative imports. No behavior change.
