+++ src/features/landing-v2/sections/FinalCTASection.jsx
@@ -5,7 +5,7 @@
 import { motion } from 'framer-motion'
 import { Loader2 } from 'lucide-react'
 
-import { useGoogleAuth } from '@/features/landing-v2/utils/useGoogleAuth'
+import { useGoogleAuth } from '@/shared/hooks/useGoogleAuth'
 import { TONE_COPY } from '@/features/landing-v2/data'
 import Button from '@/shared/ui/Button'
