+++ src/shared/lib/curatedLists.js
